requirements.txt
django
django-cors-headers
djangorestframework
django-db-connection-pool[mysql]
mysqlclient
python-dateutil
boto3
django-filter
DateTime
djangorestframework-simplejwt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
